+++ RELEASE_NOTES.md
@@ -10,6 +10,7 @@ A new header is inserted each time a *tag* is created.
 - Desktop and Web Suzanne samples now use KTX2 instead of KTX1.
 - Removed KTX1 compression support from mipgen.  [⚠️ **API Change**].
 - Added support for KTX2 to mipgen.
+- gltfio: Added resetBoneMatrices() method.
 
 ## v1.21.2